+++ lib/prouterd/iface.rb
@@ -11,6 +11,7 @@ module Prouterd
     autoload :HttpCaller,     File.expand_path("iface/http_caller",     __dir__)
     autoload :LlmCaller,      File.expand_path("iface/llm_caller",      __dir__)
     autoload :LlmSubprocess,  File.expand_path("iface/llm_subprocess",  __dir__)
+    autoload :LlmAgentic,     File.expand_path("iface/llm_agentic",     __dir__)
     autoload :PostgresCaller, File.expand_path("iface/postgres_caller", __dir__)
     autoload :LocalRepoCaller, File.expand_path("iface/local_repo_caller", __dir__)
   end